readthis integration

readthis package may be used to easily read the data from some popular mutation callers into R environment. readthis functions can be supplied not only with the single file paths, but also with lists of files or even paths to the directories with files to be loaded (and cevodata object is to store the data from many samples!)

readthis functions return tibbles or list of tibbles. These tibbles/ objects usually are instances of cevo_<software_name> S3 classes. cevomod implements methods that allow to add these types of data to the cevodata objects conveniently.

Usage

add_data(cd, data, ...)

add_to_cevodata(data, cd, name, verbose, ...)

Arguments

cd

object

data

Object read with readthis functions

...

Other arguments

name

Name for the data

verbose

Verbose?

Functions

add_data(): add_data() function takes cevodata as the first argument, so it is a preferred method for adding data in R pipelines.

add_to_cevodata(): add_to_cevodata() is a generic with a set of methods for different classes of data. These methods are called by add_data() function.

Examples

# library(cevomod)

ascat_dir <- system.file("extdata", "ASCAT", package = "readthis")
ascat <- readthis::read_ascat_files(ascat_dir)
cd <- init_cevodata("Test dataset") |>
  add_data(ascat)
#> Using 'ascat_purity' as default purity measure